+++ pyccl/halos/hbias/__init__.py
@@ -1,0 +1,14 @@
+from .hbias_base import HaloBias, halo_bias_from_name
+from .bhattacharya11 import HaloBiasBhattacharya11
+from .sheth01 import HaloBiasSheth01
+from .sheth99 import HaloBiasSheth99
+from .tinker10 import HaloBiasTinker10
+
+
+__all__ = (
+    "HaloBias", "halo_bias_from_name",
+    "HaloBiasBhattacharya11",
+    "HaloBiasSheth01",
+    "HaloBiasSheth99",
+    "HaloBiasTinker10",
+)
