+++ pyccl/halos/hmfunc/__init__.py
@@ -1,0 +1,24 @@
+from .hmfunc_base import MassFunc, mass_function_from_name
+from .angulo12 import MassFuncAngulo12
+from .bocquet16 import MassFuncBocquet16
+from .despali16 import MassFuncDespali16
+from .jenkins01 import MassFuncJenkins01
+from .press74 import MassFuncPress74
+from .sheth99 import MassFuncSheth99
+from .tinker08 import MassFuncTinker08
+from .tinker10 import MassFuncTinker10
+from .watson13 import MassFuncWatson13
+
+
+__all__ = (
+    "MassFunc", "mass_function_from_name",
+    "MassFuncAngulo12",
+    "MassFuncBocquet16",
+    "MassFuncDespali16",
+    "MassFuncJenkins01",
+    "MassFuncPress74",
+    "MassFuncSheth99",
+    "MassFuncTinker08",
+    "MassFuncTinker10",
+    "MassFuncWatson13",
+)
