+++ pyccl/halos/hmfunc/tinker08.py
@@ -1,0 +1,64 @@
+from ..massdef import MassDef200m
+from .hmfunc_base import MassFunc
+import numpy as np
+
+
+__all__ = ("MassFuncTinker08",)
+
+
+class MassFuncTinker08(MassFunc):
+    """ Implements mass function described in arXiv:0803.2706.
+
+    Args:
+        cosmo (:class:`~pyccl.core.Cosmology`): A Cosmology object.
+        mass_def (:class:`~pyccl.halos.massdef.MassDef`):
+            a mass definition object.
+            this parametrization accepts SO masses with
+            200 < Delta < 3200 with respect to the matter density.
+            If `None`, Delta = 200 (matter) will be used.
+        mass_def_strict (bool): if False, consistency of the mass
+            definition will be ignored.
+    """
+    name = 'Tinker08'
+
+    def __init__(self, cosmo, mass_def=None, mass_def_strict=True):
+        super(MassFuncTinker08, self).__init__(cosmo,
+                                               mass_def,
+                                               mass_def_strict)
+
+    def _default_mdef(self):
+        self.mdef = MassDef200m()
+
+    def _pd(self, ld):
+        return 10.**(-(0.75/(ld - 1.8750612633))**1.2)
+
+    def _setup(self, cosmo):
+        from scipy.interpolate import interp1d
+
+        delta = np.array([200.0, 300.0, 400.0, 600.0, 800.0,
+                          1200.0, 1600.0, 2400.0, 3200.0])
+        alpha = np.array([0.186, 0.200, 0.212, 0.218, 0.248,
+                          0.255, 0.260, 0.260, 0.260])
+        beta = np.array([1.47, 1.52, 1.56, 1.61, 1.87,
+                         2.13, 2.30, 2.53, 2.66])
+        gamma = np.array([2.57, 2.25, 2.05, 1.87, 1.59,
+                          1.51, 1.46, 1.44, 1.41])
+        phi = np.array([1.19, 1.27, 1.34, 1.45, 1.58,
+                        1.80, 1.97, 2.24, 2.44])
+        ldelta = np.log10(delta)
+        self.pA0 = interp1d(ldelta, alpha)
+        self.pa0 = interp1d(ldelta, beta)
+        self.pb0 = interp1d(ldelta, gamma)
+        self.pc = interp1d(ldelta, phi)
+
+    def _check_mdef_strict(self, mdef):
+        if mdef.Delta == 'fof':
+            return True
+        return False
+
+    def _get_fsigma(self, cosmo, sigM, a, lnM):
+        ld = np.log10(self._get_Delta_m(cosmo, a))
+        pA = self.pA0(ld) * a**0.14
+        pa = self.pa0(ld) * a**0.06
+        pb = self.pb0(ld) * a**self._pd(ld)
+        return pA * ((pb / sigM)**pa + 1) * np.exp(-self.pc(ld)/sigM**2)
