+++ pyccl/halos/hmfunc/tinker10.py
@@ -1,0 +1,88 @@
+from ..massdef import MassDef200m
+from .hmfunc_base import MassFunc
+import numpy as np
+
+
+__all__ = ("MassFuncTinker10",)
+
+
+class MassFuncTinker10(MassFunc):
+    """ Implements mass function described in arXiv:1001.3162.
+
+    Args:
+        cosmo (:class:`~pyccl.core.Cosmology`): A Cosmology object.
+        mass_def (:class:`~pyccl.halos.massdef.MassDef`):
+            a mass definition object.
+            this parametrization accepts SO masses with
+            200 < Delta < 3200 with respect to the matter density.
+            If `None`, Delta = 200 (matter) will be used.
+        mass_def_strict (bool): if False, consistency of the mass
+            definition will be ignored.
+        norm_all_z (bool): should we normalize the mass function
+            at z=0 or at all z?
+    """
+    __repr_attrs__ = ("mdef", "mass_def_strict", "norm_all_z",)
+    name = 'Tinker10'
+
+    def __init__(self, cosmo, mass_def=None, mass_def_strict=True,
+                 norm_all_z=False):
+        self.norm_all_z = norm_all_z
+        super(MassFuncTinker10, self).__init__(cosmo,
+                                               mass_def,
+                                               mass_def_strict)
+
+    def _default_mdef(self):
+        self.mdef = MassDef200m()
+
+    def _setup(self, cosmo):
+        from scipy.interpolate import interp1d
+
+        delta = np.array([200.0, 300.0, 400.0, 600.0, 800.0,
+                          1200.0, 1600.0, 2400.0, 3200.0])
+        alpha = np.array([0.368, 0.363, 0.385, 0.389, 0.393,
+                          0.365, 0.379, 0.355, 0.327])
+        beta = np.array([0.589, 0.585, 0.544, 0.543, 0.564,
+                         0.623, 0.637, 0.673, 0.702])
+        gamma = np.array([0.864, 0.922, 0.987, 1.09, 1.20,
+                          1.34, 1.50, 1.68, 1.81])
+        phi = np.array([-0.729, -0.789, -0.910, -1.05, -1.20,
+                        -1.26, -1.45, -1.50, -1.49])
+        eta = np.array([-0.243, -0.261, -0.261, -0.273, -0.278,
+                        -0.301, -0.301, -0.319, -0.336])
+
+        ldelta = np.log10(delta)
+        self.pA0 = interp1d(ldelta, alpha)
+        self.pa0 = interp1d(ldelta, eta)
+        self.pb0 = interp1d(ldelta, beta)
+        self.pc0 = interp1d(ldelta, gamma)
+        self.pd0 = interp1d(ldelta, phi)
+        if self.norm_all_z:
+            p = np.array([-0.158, -0.195, -0.213, -0.254, -0.281,
+                          -0.349, -0.367, -0.435, -0.504])
+            q = np.array([0.0128, 0.0128, 0.0143, 0.0154, 0.0172,
+                          0.0174, 0.0199, 0.0203, 0.0205])
+            self.pp0 = interp1d(ldelta, p)
+            self.pq0 = interp1d(ldelta, q)
+
+    def _check_mdef_strict(self, mdef):
+        if mdef.Delta == 'fof':
+            return True
+        return False
+
+    def _get_fsigma(self, cosmo, sigM, a, lnM):
+        ld = np.log10(self._get_Delta_m(cosmo, a))
+        nu = 1.686 / sigM
+        # redshift evolution only up to z=3
+        a = np.clip(a, 0.25, 1)
+        pa = self.pa0(ld) * a**(-0.27)
+        pb = self.pb0(ld) * a**(-0.20)
+        pc = self.pc0(ld) * a**0.01
+        pd = self.pd0(ld) * a**0.08
+        pA0 = self.pA0(ld)
+        if self.norm_all_z:
+            z = 1./a - 1
+            pp = self.pp0(ld)
+            pq = self.pq0(ld)
+            pA0 *= np.exp(z*(pp+pq*z))
+        return nu * pA0 * (1 + (pb * nu)**(-2 * pd)) * \
+            nu**(2 * pa) * np.exp(-0.5 * pc * nu**2)
